Refactor tests/07_task_collection to reflect new tasks structure
