+++ tests/04_api/test_task_collection_api.py
@@ -521,94 +521,3 @@ async def test_logs(
     debug(out_state["data"]["log"])
     debug(out_state["data"]["info"])
     assert out_state["data"]["log"]
-
-
-async def test_logs_failed_collection(
-    db, dummy_task_package, override_settings_factory, tmp_path: Path
-):
-    """
-    GIVEN a package and its installation environment
-    WHEN the background collection is called on it and it fails
-    THEN
-        * the log of the collection is saved to the state
-        * the installation directory is removed
-    """
-
-    override_settings_factory(
-        FRACTAL_TASKS_DIR=(tmp_path / "test_logs_failed_collection")
-    )
-
-    task_pkg = _TaskCollectPip(package=dummy_task_package.as_posix())
-
-    # Extract info form the wheel package (this would be part of the endpoint)
-    _inspect_package_and_set_attributes(task_pkg)
-    debug(task_pkg)
-
-    venv_path = create_package_dir_pip(task_pkg=task_pkg)
-    collection_status = TaskCollectStatus(
-        status="pending", venv_path=venv_path, package=task_pkg.package
-    )
-    # replacing with path because of non-serializable Path
-    collection_status_dict = collection_status.sanitised_dict()
-    state = State(data=collection_status_dict)
-    db.add(state)
-    await db.commit()
-    await db.refresh(state)
-
-    task_pkg.package = "__NO_PACKAGE"
-    task_pkg.package_path = None
-
-    await background_collect_pip(
-        state_id=state.id,
-        venv_path=venv_path,
-        task_pkg=task_pkg,
-    )
-
-    await db.refresh(state)
-    debug(state)
-    assert state.data["log"]
-    assert state.data["status"] == "fail"
-    assert state.data["info"].startswith("Original error")
-    assert not venv_path.exists()
-
-
-@pytest.mark.parametrize(
-    "relative_wheel_path",
-    (
-        "dummy_pkg_1/dist/dummy_pkg_1-0.0.1-py3-none-any.whl",
-        "dummy_pkg_2/dist/dummy_PKG_2-0.0.1-py3-none-any.whl",
-    ),
-)
-async def test_unit_create_venv_install_package(
-    testdata_path: Path,
-    tmp_path: Path,
-    override_settings_factory: callable,
-    relative_wheel_path: str,
-):
-    """
-    This unit test for `_create_venv_install_package` collects tasks from two
-    local wheel files, where the second one has a non-normalized package name.
-    """
-    from fractal_server.tasks.background_operations import (
-        _create_venv_install_package,
-    )
-    from fractal_server.logger import set_logger
-
-    LOGGER_NAME = "LOGGER"
-    set_logger(LOGGER_NAME, log_file_path=(tmp_path / "logs"))
-
-    task_package = (
-        testdata_path / "more_dummy_task_packages" / relative_wheel_path
-    )
-    task_pkg = _TaskCollectPip(package=task_package.as_posix())
-
-    # Extract info form the wheel package (this would be part of the endpoint)
-    _inspect_package_and_set_attributes(task_pkg)
-    debug(task_pkg)
-
-    # Collect task package
-    python_bin, package_root = await _create_venv_install_package(
-        task_pkg=task_pkg, path=tmp_path, logger_name=LOGGER_NAME
-    )
-    debug(python_bin)
-    debug(package_root)
