--- tests/07_task_collection/test_unit_task_utils.py
@@ -1,33 +1,0 @@
-from pathlib import Path
-
-from fractal_server.tasks.utils import _normalize_package_name
-from fractal_server.tasks.utils import get_absolute_venv_path
-
-
-def test_get_absolute_venv_path(tmp_path, override_settings_factory):
-    FRACTAL_TASKS_DIR = tmp_path / "TASKS"
-    override_settings_factory(FRACTAL_TASKS_DIR=FRACTAL_TASKS_DIR)
-    absolute_path = tmp_path
-    relative_path = Path("somewhere/else/")
-    assert get_absolute_venv_path(absolute_path) == absolute_path
-    assert get_absolute_venv_path(relative_path) == (
-        FRACTAL_TASKS_DIR / relative_path
-    )
-
-
-def test_normalize_package_name():
-    """
-    Test based on the example in
-    https://packaging.python.org/en/latest/specifications/name-normalization.
-    """
-    inputs = (
-        "friendly-bard",
-        "Friendly-Bard",
-        "FRIENDLY-BARD",
-        "friendly.bard",
-        "friendly_bard",
-        "friendly--bard",
-        "FrIeNdLy-._.-bArD",
-    )
-    outputs = list(map(_normalize_package_name, inputs))
-    assert len(set(outputs)) == 1
